# app/api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: bioalign API
# basePath prefixes all resource paths 
basePath: /api/v1
# 
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /alignments:
    x-swagger-router-controller: alignments
    post:
      tags : [ Alignment ]
      description: Request a new asynchronous alignment
      operationId: requestAlignment
      parameters:
        - name: request
          in: body
          description: Alignment specification
          required: true
          schema:
            $ref: "#/definitions/AlignRequest"
      responses:
        "202":
          description: Success
          schema:
            $ref: "#/definitions/AlignTaskStatus"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      tags : [ Alignment ]
      description: Get completed alignments
      operationId: getAlignments
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AlignResponseCollection"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /alignments/{id}:
    x-swagger-router-controller: alignments
    get:
      tags : [ Alignment ]
      description: Get durable result of specific completed alignment
      operationId: getAlignment
      parameters:
        - name: id
          in: path
          description: Alignment id
          required: true
          type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AlignResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /tasks/{taskId}:
    x-swagger-router-controller: tasks
    get:
      tags : [ Alignment ]
      description: Get status of an ongoing or recent alignment
      operationId: requestStatus
      parameters:
        - name: taskId
          in: path
          description: Task identifier
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/AlignTaskStatus"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /health:
    x-swagger-router-controller: health
    get:
      tags : [ Health ]
      description: Check health
      operationId: checkHealth
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/HealthResponse"
definitions:
  AlignRequest:
    required:
      - sequence
    properties:
      sequence:
        type: string
        example: AATCCGCTAG
  AlignTaskStatus:
    properties:
      taskId:
        type: string
      status:
        type: string
      message:
        type: string
  AlignResponse:
    properties:
      date:
        type: string
      time:
        type: string
      sequence:
        type: string
      proteinName:
        type: string
      proteinPosition:
        type: number
  AlignResponseCollection:
    properties:
      count:
        type: integer
      items:
        type: array
        items:
          $ref: "#/definitions/AlignResponse"
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  HealthResponse:
    type: string